tets

+++ Task8/Task8.cpp
@@ -6,7 +6,7 @@
 #include "omp.h"
 #include "Task8.h"
 
-void Task8::execute(int threads_amount, int N) {
+void Task8::execute(int threads_amount, long N) {
     printf("\n\nInput: N = %d \nOutput: \n", N);
 
     omp_set_num_threads(threads_amount);
